pyproject.toml
[tool.poetry]
name = "c-block-backend"
version = "0.1.0"
description = ""
authors = ["Euphemia Deathridge <ephdtrg@mintyclouds.in>"]

[tool.poetry.dependencies]
python = "^3.8"
web3 = "^5.17.0"
Django = "^3.2.11"
djangorestframework = "^3.12.4"
drf-yasg = "^1.20.0"
psycopg2-binary = "^2.9.2"
marshmallow = "^3.14.1"
marshmallow-dataclass = "^8.5.3"
marshmallow-enum = "^1.5.1"
PyYAML = "^6.0"
django-extensions = "^3.1.5"
ptpython = "^3.0.20"
pytest-django = "^4.4.0"
jsonschema = "^3.2.0"
APScheduler = "^3.8.1"
redis = "^4.0.2"
dramatiq = "^1.12.0"
django-dramatiq = "^0.10.0"
sentry-sdk = "^1.5.7"
shyaml = "^0.6.2"
gunicorn = "^20.1.0"
django-rest-auth = "^0.9.5"
django-allauth = "^0.50.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
